Test react native/django/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pillow
mysqlclient